Removed hardware-configuration.nix as it is unnecesary
This file is separated into nvidia.nix, kernel.nix, and firewall.nix. So, this file which is generated during installation can be ignored. If this is present and pulled in, the UUID of the disks will be wrong the OS won't boot. So, better not to include this file. Only the systemd-boot must be disabled or removed from the hardware-configuration.nix file generated during installation.